feat: рефакторинг и более корректное удаление поста

+++ app/Http/Controllers/Post/DeletePostController.php
@@ -2,11 +2,14 @@
 
 namespace App\Http\Controllers\Post;
 
+use App\Exceptions\Post\YouAreNotAnAuthorOfThisPostException;
 use App\Http\Controllers\Controller;
 use App\Models\Post;
 use App\Services\Post\PostService;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class DeletePostController extends Controller
 {
@@ -14,8 +17,19 @@ class DeletePostController extends Controller
         private readonly PostService $postService
     ) {}
 
+    /**
+     * @throws YouAreNotAnAuthorOfThisPostException
+     */
     public function __invoke(Request $request, Post $post): JsonResponse
     {
+        if ($post->is_deleted) {
+            throw new ModelNotFoundException();
+        }
+
+        if (!Gate::allows('edit-post', $post)) {
+            throw new YouAreNotAnAuthorOfThisPostException();
+        }
+
         $this->postService->deletePost(post: $post);
 
         return response()->json([
